backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
django-filter>=23.5
Pillow>=10.2
dj-database-url>=2.1
gunicorn>=21.2
psycopg[binary]>=3.1
aiogram>=3.3,<4.0
httpx>=0.26,<0.27
